backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-google-genai
chromadb
rank_bm25
sentence-transformers
pymupdf
pydantic
ragas
pytest
fpdf
python-dotenv
langchain-huggingface
python-multipart
pandas
scikit-learn
joblib
